fix: remove old allocation after setting new one
In _setAllocation() the old allocation was being subtracted with an
`updateMemberUnits()` call before adding the new one with the same call.
If there is only one member the total units between the two calls became
zero and the flow would stop requiring a new distribution to start
again.
This changes the order of the calls so that the units are always more
than zero (unless the member actually wants that) and removes one loop.